Move sqlite_db::rollover::delete to rollover::overwrite module
It is only used in the `overwrite` module.
