+++ sqlite-db/src/rollover/overwrite/delete.rs
